# [prototype only] Reuse existing arrays in GroupByHash