Prototype only



## The following is a benchmark comparing when we convert arrays to big arrays
# It is irrelative to the current patch; just for referrence

Primitive type arrays used in GroupByHash can grow to GB sizes. Each
rehash can discard an array of the same size that can cause humongous
allocation repeatedly until there is no need to grow capacity. We found
simple queries in production (e.g., `SELECT DISTINCT a from t`) can cause
full GCs. This patch replaces all primitive type arrays with big arrays
to avoid humongous allocation.

## pretty severe regression; need to revise:
Random access benchmark wrt two dimensional arrays vs one dimensional arrays (https://stackoverflow.com/questions/2512082/java-multi-dimensional-array-vs-one-dimensional)
## laptop:
**before:**
```
Benchmark                                   (channelCount)  (groupCount)  (hashEnabled)  Mode  Cnt     Score     Error  Units
BenchmarkGroupByHash.addPagePreCompute                   1       3000000           true  avgt   20   219.102 ±  11.709  ns/op
BenchmarkGroupByHash.addPagePreCompute                   1       3000000          false  avgt   20   226.143 ±  13.646  ns/op
BenchmarkGroupByHash.addPagePreCompute                  10       3000000           true  avgt   20   789.242 ±  80.450  ns/op
BenchmarkGroupByHash.addPagePreCompute                  10       3000000          false  avgt   20  1168.223 ±  71.983  ns/op
BenchmarkGroupByHash.addPagePreCompute                  20       3000000           true  avgt   20  1590.167 ± 135.052  ns/op
BenchmarkGroupByHash.addPagePreCompute                  20       3000000          false  avgt   20  2189.896 ±  76.140  ns/op
BenchmarkGroupByHash.groupByHashPreCompute               1       3000000           true  avgt   20   196.256 ±  11.436  ns/op
BenchmarkGroupByHash.groupByHashPreCompute               1       3000000          false  avgt   20   239.136 ±   7.145  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              10       3000000           true  avgt   20   641.710 ±  14.017  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              10       3000000          false  avgt   20  1109.391 ±  52.217  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              20       3000000           true  avgt   20  1658.992 ± 178.336  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              20       3000000          false  avgt   20  2234.991 ± 101.736  ns/op
```
**after:**
```
Benchmark                                   (channelCount)  (groupCount)  (hashEnabled)  Mode  Cnt     Score     Error  Units
BenchmarkGroupByHash.addPagePreCompute                   1       3000000           true  avgt   10   224.806 ±  23.703  ns/op
BenchmarkGroupByHash.addPagePreCompute                   1       3000000          false  avgt   10   250.339 ±   3.986  ns/op
BenchmarkGroupByHash.addPagePreCompute                  10       3000000           true  avgt   10   757.254 ± 106.717  ns/op
BenchmarkGroupByHash.addPagePreCompute                  10       3000000          false  avgt   10  1181.979 ±  76.400  ns/op
BenchmarkGroupByHash.addPagePreCompute                  20       3000000           true  avgt   10  1722.848 ± 308.191  ns/op
BenchmarkGroupByHash.addPagePreCompute                  20       3000000          false  avgt   10  2510.248 ± 443.145  ns/op
BenchmarkGroupByHash.groupByHashPreCompute               1       3000000           true  avgt   10   248.502 ±  35.051  ns/op
BenchmarkGroupByHash.groupByHashPreCompute               1       3000000          false  avgt   10   289.508 ±  35.033  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              10       3000000           true  avgt   10   757.911 ±  82.060  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              10       3000000          false  avgt   10  1468.663 ± 202.592  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              20       3000000           true  avgt   10  2047.919 ± 311.392  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              20       3000000          false  avgt   10  3018.833 ± 613.826  ns/op
```

## Server:
**before:**
```
Benchmark                                   (channelCount)  (groupCount)  (hashEnabled)  Mode  Cnt     Score     Error  Units
BenchmarkGroupByHash.addPagePreCompute                   1       3000000           true  avgt   40   252.108 ±  15.141  ns/op
BenchmarkGroupByHash.addPagePreCompute                   1       3000000          false  avgt   40   278.329 ±  17.196  ns/op
BenchmarkGroupByHash.addPagePreCompute                  10       3000000           true  avgt   40   999.718 ±  42.933  ns/op
BenchmarkGroupByHash.addPagePreCompute                  10       3000000          false  avgt   40  1654.101 ± 108.303  ns/op
BenchmarkGroupByHash.addPagePreCompute                  20       3000000           true  avgt   40  2227.519 ±  92.364  ns/op
BenchmarkGroupByHash.addPagePreCompute                  20       3000000          false  avgt   40  3347.404 ± 129.901  ns/op
BenchmarkGroupByHash.groupByHashPreCompute               1       3000000           true  avgt   40   304.079 ±  18.457  ns/op
BenchmarkGroupByHash.groupByHashPreCompute               1       3000000          false  avgt   40   299.655 ±  16.705  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              10       3000000           true  avgt   40  1013.963 ±  45.575  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              10       3000000          false  avgt   40  1713.749 ±  82.685  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              20       3000000           true  avgt   40  2305.849 ± 104.726  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              20       3000000          false  avgt   40  3493.750 ± 141.248  ns/op
```

**after (`long[]` => `LongBigArray`, `int[]` => `IntBigArray`, and `byte[]` => `ByteBigArray`):**
```
Benchmark                                   (channelCount)  (groupCount)  (hashEnabled)  Mode  Cnt     Score     Error  Units
BenchmarkGroupByHash.addPagePreCompute                   1       3000000           true  avgt   40   312.157 ±  13.091  ns/op
BenchmarkGroupByHash.addPagePreCompute                   1       3000000          false  avgt   40   352.682 ±  19.351  ns/op
BenchmarkGroupByHash.addPagePreCompute                  10       3000000           true  avgt   40  1039.557 ±  44.293  ns/op
BenchmarkGroupByHash.addPagePreCompute                  10       3000000          false  avgt   40  1778.058 ±  78.182  ns/op
BenchmarkGroupByHash.addPagePreCompute                  20       3000000           true  avgt   40  2363.148 ±  92.150  ns/op
BenchmarkGroupByHash.addPagePreCompute                  20       3000000          false  avgt   40  3559.444 ± 146.411  ns/op
BenchmarkGroupByHash.groupByHashPreCompute               1       3000000           true  avgt   40   338.067 ±  18.628  ns/op
BenchmarkGroupByHash.groupByHashPreCompute               1       3000000          false  avgt   40   379.388 ±  14.095  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              10       3000000           true  avgt   40  1144.029 ±  22.583  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              10       3000000          false  avgt   40  1682.492 ±  92.457  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              20       3000000           true  avgt   40  2436.306 ±  79.558  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              20       3000000          false  avgt   40  3617.578 ± 119.018  ns/op
```
**after (`long[]` remains unchanged, `int[]` => `IntBigArray`, and `byte[]` => `ByteBigArray`):**
```
Benchmark                                   (channelCount)  (groupCount)  (hashEnabled)  Mode  Cnt     Score     Error  Units
BenchmarkGroupByHash.addPagePreCompute                   1       3000000           true  avgt   40   290.396 ±  18.097  ns/op
BenchmarkGroupByHash.addPagePreCompute                   1       3000000          false  avgt   40   320.971 ±  19.822  ns/op
BenchmarkGroupByHash.addPagePreCompute                  10       3000000           true  avgt   40  1047.571 ±  41.309  ns/op
BenchmarkGroupByHash.addPagePreCompute                  10       3000000          false  avgt   40  1791.248 ±  81.908  ns/op
BenchmarkGroupByHash.addPagePreCompute                  20       3000000           true  avgt   40  2267.405 ±  85.051  ns/op
BenchmarkGroupByHash.addPagePreCompute                  20       3000000          false  avgt   40  3569.935 ± 120.969  ns/op
BenchmarkGroupByHash.groupByHashPreCompute               1       3000000           true  avgt   40   313.321 ±  20.825  ns/op
BenchmarkGroupByHash.groupByHashPreCompute               1       3000000          false  avgt   40   335.576 ±  16.511  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              10       3000000           true  avgt   40  1048.885 ±  39.247  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              10       3000000          false  avgt   40  1708.691 ±  70.837  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              20       3000000           true  avgt   40  2294.402 ± 100.122  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              20       3000000          false  avgt   40  3478.707 ± 113.066  ns/op
```
**after (`long[]` remains unchanged, `int[]` remains unchanged, and `byte[]` => `ByteBigArray`):**
```
Benchmark                                   (channelCount)  (groupCount)  (hashEnabled)  Mode  Cnt     Score     Error  Units
BenchmarkGroupByHash.addPagePreCompute                   1       3000000           true  avgt   40   268.928 ±  15.794  ns/op
BenchmarkGroupByHash.addPagePreCompute                   1       3000000          false  avgt   40   303.769 ±  19.028  ns/op
BenchmarkGroupByHash.addPagePreCompute                  10       3000000           true  avgt   40  1041.915 ±  52.588  ns/op
BenchmarkGroupByHash.addPagePreCompute                  10       3000000          false  avgt   40  1689.601 ±  62.508  ns/op
BenchmarkGroupByHash.addPagePreCompute                  20       3000000           true  avgt   40  2206.633 ± 108.120  ns/op
BenchmarkGroupByHash.addPagePreCompute                  20       3000000          false  avgt   40  3588.593 ± 168.011  ns/op
BenchmarkGroupByHash.groupByHashPreCompute               1       3000000           true  avgt   40   321.344 ±  18.841  ns/op
BenchmarkGroupByHash.groupByHashPreCompute               1       3000000          false  avgt   40   346.114 ±  23.600  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              10       3000000           true  avgt   40  1081.986 ±  39.844  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              10       3000000          false  avgt   40  1765.118 ±  83.795  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              20       3000000           true  avgt   40  2260.910 ± 105.671  ns/op
BenchmarkGroupByHash.groupByHashPreCompute              20       3000000          false  avgt   40  3575.890 ± 135.284  ns/op
```