@highker 2017-11-10
Haven't seen this issue happening in production. Close for now.